% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{dnormmix}
\alias{dnormmix}
\title{Compute the PDF of a normal mixture distribution}
\usage{
dnormmix(x, mean_1, sd_1, mean_2, sd_2, mixprob)
}
\arguments{
\item{x}{Numeric value(s) where the PDF is evaluated.}

\item{mean_1}{Mean of the first normal distribution.}

\item{sd_1}{Standard deviation of the first normal distribution.}

\item{mean_2}{Mean of the second normal distribution.}

\item{sd_2}{Standard deviation of the second normal distribution.}

\item{mixprob}{Probability weight for the first normal distribution (between 0 and 1).}
}
\value{
A numeric value representing the PDF at `x`.
}
\description{
Compute the PDF of a normal mixture distribution
}
