% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{rnormmix}
\alias{rnormmix}
\title{Generate random numbers from a normal mixture distribution}
\usage{
rnormmix(n, mean_1, sd_1, mean_2, sd_2, mixprob)
}
\arguments{
\item{n}{Number of random samples to generate.}

\item{mean_1}{Mean of the first normal distribution.}

\item{sd_1}{Standard deviation of the first normal distribution.}

\item{mean_2}{Mean of the second normal distribution.}

\item{sd_2}{Standard deviation of the second normal distribution.}

\item{mixprob}{Probability weight for the first normal distribution (between 0 and 1).}
}
\value{
A vector of `n` random values from the mixture distribution.
}
\description{
Generate random numbers from a normal mixture distribution
}
